Refactor ExcludePostsControl and PostIncludesControl
- both now use a common PostPickerControl
- legacy support for ExcludePostsControl where the block attribute is an array of IDs
