+++ includes/Traits/Exclude_Posts.php
@@ -28,6 +28,15 @@ trait Exclude_Posts {
 		// If there are already posts to be excluded, we need to add to them.
 		$exclude_ids = $this->custom_args['post__not_in'] ?? array();
 
+		if ( empty( $to_exclude ) ) {
+			return $exclude_ids;
+		}
+
+		if ( is_numeric( $to_exclude ) ) {
+			$to_exclude = [ $to_exclude ];
+		} elseif ( is_array( $to_exclude ) && isset( $to_exclude[0]['id'] ) ) {
+			$to_exclude = array_column( $to_exclude, 'id' );
+		}
 		$exclude_ids = array_unique( array_merge( $exclude_ids, (array) $to_exclude ) );
 
 		return $exclude_ids;
